Create recipe fullplate

+++ src/main/java/com/cookplanner/repositories/Roles.java
@@ -10,7 +10,7 @@ import java.util.List;
  * @author Created by Yassine on 2/19/17.
  */
 @Repository
-public interface Roles extends CrudRepository<UserRole, Long> {
+public interface Roles extends CrudRepository<UserRole, Integer> {
     @Query("select ur.role from UserRole ur, User u where u.username=?1 and ur.userId = u.id")
     public List<String> ofUserWith (String username);
